Fix getHumanHorizon when the jet banks to the right
The getHumanHorizonFromPitchRollAzEl function seems to work fine during Gimbal. In other situations where the jet is banked to the right the function can result in the wrong sign for the angle of the horizon.
The angleTo function that it uses always returns a positive value. The fix is to return a signed angle instead relative to the az/el heading.
This currently has no effect on Sitrec, but the function could be used by scripts for other purposes.