Register nodes needed for a sitch in console mode
In console mode most nodes and keys in the sitch data cannot and/or do not need to be imported and used.
This provides a way to opt-in to which ones should be registered and used.
With these changes almost all of the calculations in Gimbal/GoFast should be accessible.
Additional nodes/keys can be enabled later as necessary for other sitches.

Also:
- fixed some issues with certain nodes in console mode
- added a test for running GoFast in console mode